Move bin+onboard into new shipit namespace
