+++ components/nav.js
@@ -143,7 +143,7 @@ const Navigation = props => (
     </NextLink>
     <Link href="https://scrapbook.hackclub.com/">Scrapbook</Link>
     <Link href="https://toolbox.hackclub.com/">Toolbox</Link>
-    <NextLink href="/onboard" passHref>
+    <NextLink href="/shipit/onboard" passHref>
       <Link>OnBoard</Link>
     </NextLink>
   </NavBar>
