+++ pages/shipit/onboard/index.js
@@ -2,17 +2,17 @@ import { Box, Button, Flex, Grid, Heading, Image, Link, Text } from 'theme-ui'
 import Balancer from 'react-wrap-balancer'
 import Head from 'next/head'
 import Meta from '@hackclub/meta'
-import Nav from '../../components/nav'
-import Footer from '../../components/footer'
-import FadeIn from '../../components/fade-in'
-import Sparkles from '../../components/sparkles'
-import Tilt from '../../components/tilt'
-import Recap from '../../components/onboard/recap'
-import usePrefersReducedMotion from '../../lib/use-prefers-reduced-motion'
+import Nav from '../../../components/nav'
+import Footer from '../../../components/footer'
+import FadeIn from '../../../components/fade-in'
+import Sparkles from '../../../components/sparkles'
+import Tilt from '../../../components/tilt'
+import Recap from '../../../components/onboard/recap'
+import usePrefersReducedMotion from '../../../lib/use-prefers-reduced-motion'
 import { useEffect, useRef, useState } from 'react'
-import sleep from '../../lib/sleep'
-import Announcement from '../../components/announcement'
-import YoutubeVideo from '../../components/onboard/youtube-video'
+import sleep from '../../../lib/sleep'
+import Announcement from '../../../components/announcement'
+import YoutubeVideo from '../../../components/onboard/youtube-video'
 import Icon from '@hackclub/icons'
 
 /**
